[INTEL_HPU] seperated UT cast into w/o 8k/4K
seperated UT cast test cases into two suites:
test_cast.py only needs 11s, this will be included
into PR testing
test_cast_8k_4k.py will needs about 10mins, this will
be run on daily regression testing
